# spec/api/api.yaml
openapi: 3.0.0
info:
  title: HoldYouBeer API
  description: |
    The official API for the HoldYourBeer application, serving both the iOS and Web clients.

    **API Versioning**: This API uses URL-based versioning. All endpoints are prefixed with `/api/v1/`.

    **Note**: Legacy non-versioned endpoints (e.g., `/api/beers`) are deprecated and will be removed on 2026-12-31.
    Please migrate to versioned endpoints (e.g., `/api/v1/beers`).
  version: 1.1.0

paths:
  # ... Authentication endpoints remain the same ...
  /api/v1/register:
    post:
      summary: Register a new user
      tags:
        - Authentication
      # ... (content omitted for brevity)

  /api/v1/login:
    post:
      summary: Request an API token (Login)
      tags:
        - Authentication
      # ... (content omitted for brevity)

  /api/v1/auth/firebase/login:
    post:
      summary: Authenticate with Firebase ID Token
      description: |
        Authenticate a user using a Firebase ID Token obtained from Firebase Auth on the client side.
        Supports Google Sign-In, Apple Sign-In, and other Firebase authentication methods.
        Creates a new user account if the Firebase UID doesn't exist in the database.
      tags:
        - Firebase Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id_token
              properties:
                id_token:
                  type: string
                  description: Firebase ID Token obtained from Firebase Auth SDK
                  example: "eyJhbGciOiJSUzI1NiIsImtpZCI6IjFkYzBmMTc..."
                fcm_token:
                  type: string
                  description: Optional FCM token for push notifications
                  example: "fGcm:token:example..."
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Login successful"
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      name:
                        type: string
                        example: "John Doe"
                      email:
                        type: string
                        example: "john@example.com"
                      provider:
                        type: string
                        enum: [google, apple, email, firebase]
                        example: "google"
                      firebase_uid:
                        type: string
                        example: "firebase-uid-123"
        '401':
          description: Invalid or expired Firebase token
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Unauthorized"
                  message:
                    type: string
                    example: "Invalid or expired Firebase token"
        '500':
          description: Authentication failed due to server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Authentication failed"
                  message:
                    type: string
                    example: "An error occurred during authentication"

  /api/v1/auth/firebase/me:
    get:
      summary: Get authenticated user information
      description: Returns the current user's profile information. Requires a valid Firebase ID Token.
      tags:
        - Firebase Authentication
      security:
        - firebaseAuth: []
      responses:
        '200':
          description: User information retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      name:
                        type: string
                        example: "John Doe"
                      email:
                        type: string
                        example: "john@example.com"
                      provider:
                        type: string
                        example: "google"
                      firebase_uid:
                        type: string
                        example: "firebase-uid-123"
                      email_verified_at:
                        type: string
                        format: date-time
                        example: "2025-11-05T10:30:45+00:00"
                      created_at:
                        type: string
                        format: date-time
                        example: "2025-11-05T10:30:45+00:00"
        '401':
          description: Unauthorized - Invalid or missing token

  /api/v1/auth/firebase/fcm-token:
    post:
      summary: Update FCM token for push notifications
      description: Updates the user's FCM token used for sending push notifications.
      tags:
        - Firebase Authentication
      security:
        - firebaseAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - fcm_token
              properties:
                fcm_token:
                  type: string
                  description: FCM token from Firebase Cloud Messaging
                  example: "fGcm:token:example..."
      responses:
        '200':
          description: FCM token updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "FCM token updated successfully"
        '401':
          description: Unauthorized - Invalid or missing token

  /api/v1/auth/firebase/logout:
    post:
      summary: Logout user
      description: Clears the user's FCM token. The client should also sign out from Firebase Auth.
      tags:
        - Firebase Authentication
      security:
        - firebaseAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Logout successful"
        '401':
          description: Unauthorized - Invalid or missing token

  # ... Beer and Brand Endpoints ...
  /api/v1/beers:
    get:
      summary: Get my list of tracked beers
      description: Returns a paginated list of all unique beers the user has tracked, including the brand and how many times they've been consumed.
      tags:
        - Beers
      security:
        - bearerAuth: []
      parameters:
        - name: per_page
          in: query
          description: "Number of items per page. Default is 20, maximum is 100."
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
            example: 20
        - name: page
          in: query
          description: "Page number. Default is 1."
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            example: 1
        - name: sort
          in: query
          description: "Sort order. Options: '-tasted_at' (newest first, default), 'tasted_at' (oldest first), 'name' (A-Z), '-name' (Z-A)."
          required: false
          schema:
            type: string
            enum: ["-tasted_at", "tasted_at", "name", "-name"]
            default: "-tasted_at"
            example: "-tasted_at"
        - name: brand_id
          in: query
          description: "Filter beers by a specific brand ID. Useful for series name suggestions."
          required: false
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: A paginated list of my beers with metadata.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Beer'
                  meta:
                    type: object
                    properties:
                      current_page:
                        type: integer
                        example: 1
                      last_page:
                        type: integer
                        example: 5
                      per_page:
                        type: integer
                        example: 20
                      total:
                        type: integer
                        example: 95
                      from:
                        type: integer
                        example: 1
                      to:
                        type: integer
                        example: 20
                  links:
                    type: object
                    properties:
                      first:
                        type: string
                        example: "http://localhost/api/v1/beers?page=1"
                      last:
                        type: string
                        example: "http://localhost/api/v1/beers?page=5"
                      prev:
                        type: string
                        nullable: true
                        example: null
                      next:
                        type: string
                        nullable: true
                        example: "http://localhost/api/v1/beers?page=2"
    post:
      summary: Add a new beer to track
      description: Adds a new unique beer, potentially creating a new brand if it doesn't exist.
      tags:
        - Beers
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                brand_name:
                  type: string
                  example: "Guinness"
                name:
                  type: string
                  example: "Draught"
                style:
                  type: string
                  example: "Stout"
      responses:
        '201':
          description: The new beer was created and added to the user's tracked list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Beer'

  /api/v1/beers/{id}/count_actions:
    post:
      summary: Increment or decrement the tasting count for a beer
      description: The primary endpoint for modifying a beer count. The backend handles updating the count table and creating a log entry in a database transaction with row-level locking (lockForUpdate) to ensure data consistency and prevent race conditions.
      tags:
        - Beers
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
                  enum: [increment, decrement]
                  example: "increment"
      responses:
        '200':
          description: Action was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Beer' # Return the updated beer object

  /api/v1/beers/{id}/tasting_logs:
    get:
      summary: Get the tasting log for a specific beer
      description: Returns a list of all tasting log events for a single beer.
      tags:
        - Beers
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: A list of tasting log events.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TastingLog'

  /api/v1/brands:
    get:
      summary: Get all brands
      tags:
        - Brands
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of all available brands.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Brand'

components:
  schemas:
    Brand:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Guinness"

    Beer:
      type: object
      properties:
        id:
          type: integer
          example: 10
        name:
          type: string
          description: The specific name or series of the beer.
          example: "Draught"
        style:
          type: string
          example: "Stout"
        brand:
          $ref: '#/components/schemas/Brand'
        tasting_count:
          type: integer
          description: How many times the current user has tasted this beer. This value is read from a dedicated count table for performance.
          example: 5
        last_tasted_at:
          type: string
          format: date-time
          description: The timestamp of when the user last tasted this beer.
          example: "2025-11-05T10:30:45+00:00"

    TastingLog:
      type: object
      properties:
        id:
          type: integer
          description: The unique ID of the log event.
          example: 101
        action:
          type: string
          enum: [increment, decrement, create]
          description: The action that was performed.
          example: "increment"
        logged_at:
          type: string
          format: date-time
          description: The timestamp of when the action was logged.
          example: "2025-08-12T21:00:00Z"

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Sanctum token authentication for traditional email/password login
    firebaseAuth:
      type: http
      scheme: bearer
      description: Firebase ID Token authentication for Google/Apple Sign-In